+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+* Duration timeout parameter added to NetworkTester methods
+
+
 ### Changed
 
 * update to JUnit 6